requirements.txt
tweepy
requests
mecab-python3
